backend/pyproject.toml
[tool.poetry]
name = "rag-chatbot-backend"
version = "0.1.0"
description = "Production-ready RAG chatbot for AAOIFI Sharia Standards"
authors = ["Talal Ahmad <heyytalall@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
langchain = "^0.1.0"
langchain-groq = "^0.0.1"
langchain-pinecone = "^0.0.4"
langchain-community = "^0.0.10"
langchain-openai = "^0.0.2"
pinecone-client = "^3.0.0"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
python-dotenv = "^1.0.0"
unstructured = {extras = ["pdf"], version = "^0.12.0"}
pymupdf = "^1.23.0"
python-multipart = "^0.0.6"
structlog = "^24.1.0"
slowapi = "^0.1.9"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
black = "^24.1.0"
ruff = "^0.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
